+++ cmd/frontend/internal/graphqlbackend/repository_text_search_index.go
@@ -112,7 +112,11 @@ func (r *repositoryTextSearchIndexResolver) Refs(ctx context.Context) ([]*reposi
 	}
 	if entry != nil {
 		for _, branch := range entry.Repository.Branches {
-			ref := refByName("refs/heads/" + branch.Name)
+			name := "refs/heads/" + branch.Name
+			if branch.Name == "HEAD" {
+				name = defaultBranchRef.name
+			}
+			ref := refByName(name)
 			ref.indexedCommit = gitObjectID(branch.Version)
 		}
 	}